power: battery_current_limit: Support to get the minimum frequency limit
BCL driver supports a phandle reference to the msm_thermal's
vdd restriction node. BCL uses the phandle to get the minimum
frequency limit that the thermal driver places during vdd
restriction. This frequency limit is the minimum value that
the BCL driver can limit the CPU max frequency to.

CRs-Fixed: 629894
Change-Id: I0a4ff50b1759e2abe97d732da8a3de71f58bc538
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>