qcom: clock-debug: Print actual voltage vote in 'enabled_clocks' output
Currently, when one dumps the list of enabled clocks in
the kernel log or in the shell, the voltage level of the
VDD class is printed. This is not really useful and the
information can already be obtained easily from the
regulator debugfs nodes. Instead, print the actual voltage
that each clock has voted for, since multiple clocks may
share the same VDD class. Otherwise the output is confusing,
because every clock sharing a VDD class will include the
same voltage vote (the current overall vote on the
regulator) even though that individual clock's vote may
actually be lower.

Change-Id: I4ec550ee404a2612323486b2000465341a5ff2e2
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>