ARM: dts: msm: Add pwm-cells property to PWM devices
pwm-cells specifier has to be defined for the PWM devices. This is
needed for the client drivers to get PWM devices through pwm_get().
Also, it can be used by the PWM framework to set the chip specific
implementation of xlate when it's being added to the PWM device list.

Change-Id: If11eceb0e93f2b4e22f78c3b1a01f2b8f3b7247f
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>

Conflicts:
	arch/arm/boot/dts/qcom/msm-pm8994.dtsi