char: msm_smd_pkt: Reduce log severity of signals
If the user process is interrupted by a signal while the smd_pkt driver
is executing a wait on behalf of the client, this activity gets logged as
an error.  Since this scenario may be expected, and generally can be
transparent to the client as VFS may restart the syscall, the log level
of this event should be reduced.

Reduce log events related to signal handling as INFO so that they can be
filtered out as appropriate.

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
https://github.com/neobuddy89/shamu/commit/f5c1b07b46fa2bf0352854ca00efdc57e7e7dde1

Change-Id: I685fb7720da846b76c9039eefd83fd655187312f